[tests] Import pipeline/model classes inside the tests that use them so the fetcher does not over-select through shared mixins
