+++ tests/models/test_modeling_common.py
@@ -26,7 +26,7 @@ import torch
 from huggingface_hub import ModelCard, delete_repo, snapshot_download, try_to_load_from_cache
 from huggingface_hub.utils import HfHubHTTPError, is_jinja_available
 
-from diffusers.models import FluxTransformer2DModel, SD3Transformer2DModel, UNet2DConditionModel
+from diffusers.models import UNet2DConditionModel
 
 from ..others.test_utils import TOKEN, USER, is_staging_test
 from ..testing_utils import (
@@ -118,6 +118,8 @@ class TestModelUtils:
                 assert False, "Parameters not the same!"
 
     def test_local_files_only_with_sharded_checkpoint(self):
+        from diffusers.models import FluxTransformer2DModel
+
         repo_id = "hf-internal-testing/tiny-flux-sharded"
         error_response = mock.Mock(
             status_code=500,
@@ -232,6 +234,8 @@ class TestModelUtils:
         A simple tests to check if the modules under `_keep_in_fp32_modules` are kept in fp32 when we load the model in fp16/bf16
         Also ensures if inference works.
         """
+        from diffusers.models import SD3Transformer2DModel
+
         fp32_modules = SD3Transformer2DModel._keep_in_fp32_modules
 
         for torch_dtype in [torch.bfloat16, torch.float16]:
