+++ tests/pipelines/pag/test_pag_sd.py
@@ -23,7 +23,6 @@ from transformers import CLIPTextConfig, CLIPTextModel, CLIPTokenizer
 
 from diffusers import (
     AutoencoderKL,
-    AutoPipelineForText2Image,
     DDIMScheduler,
     StableDiffusionPAGPipeline,
     StableDiffusionPipeline,
@@ -315,6 +314,8 @@ class StableDiffusionPAGPipelineIntegrationTests(unittest.TestCase):
         return inputs
 
     def test_pag_cfg(self):
+        from diffusers import AutoPipelineForText2Image
+
         pipeline = AutoPipelineForText2Image.from_pretrained(self.repo_id, enable_pag=True, torch_dtype=torch.float16)
         pipeline.enable_model_cpu_offload(device=torch_device)
         pipeline.set_progress_bar_config(disable=None)
@@ -333,6 +334,8 @@ class StableDiffusionPAGPipelineIntegrationTests(unittest.TestCase):
         )
 
     def test_pag_uncond(self):
+        from diffusers import AutoPipelineForText2Image
+
         pipeline = AutoPipelineForText2Image.from_pretrained(self.repo_id, enable_pag=True, torch_dtype=torch.float16)
         pipeline.enable_model_cpu_offload(device=torch_device)
         pipeline.set_progress_bar_config(disable=None)
