+++ tests/pipelines/pag/test_pag_sd_img2img.py
@@ -25,7 +25,6 @@ from transformers import CLIPTextConfig, CLIPTextModel, CLIPTokenizer
 from diffusers import (
     AutoencoderKL,
     AutoencoderTiny,
-    AutoPipelineForImage2Image,
     EulerDiscreteScheduler,
     StableDiffusionImg2ImgPipeline,
     StableDiffusionPAGImg2ImgPipeline,
@@ -254,6 +253,8 @@ class StableDiffusionPAGImg2ImgPipelineIntegrationTests(unittest.TestCase):
         return inputs
 
     def test_pag_cfg(self):
+        from diffusers import AutoPipelineForImage2Image
+
         pipeline = AutoPipelineForImage2Image.from_pretrained(self.repo_id, enable_pag=True, torch_dtype=torch.float16)
         pipeline.enable_model_cpu_offload(device=torch_device)
         pipeline.set_progress_bar_config(disable=None)
@@ -272,6 +273,8 @@ class StableDiffusionPAGImg2ImgPipelineIntegrationTests(unittest.TestCase):
         )
 
     def test_pag_uncond(self):
+        from diffusers import AutoPipelineForImage2Image
+
         pipeline = AutoPipelineForImage2Image.from_pretrained(self.repo_id, enable_pag=True, torch_dtype=torch.float16)
         pipeline.enable_model_cpu_offload(device=torch_device)
         pipeline.set_progress_bar_config(disable=None)
