+++ tests/pipelines/pag/test_pag_sdxl_img2img.py
@@ -32,7 +32,6 @@ from transformers import (
 
 from diffusers import (
     AutoencoderKL,
-    AutoPipelineForImage2Image,
     EulerDiscreteScheduler,
     StableDiffusionXLImg2ImgPipeline,
     StableDiffusionXLPAGImg2ImgPipeline,
@@ -302,6 +301,8 @@ class StableDiffusionXLPAGImg2ImgPipelineIntegrationTests(unittest.TestCase):
         return inputs
 
     def test_pag_cfg(self):
+        from diffusers import AutoPipelineForImage2Image
+
         pipeline = AutoPipelineForImage2Image.from_pretrained(self.repo_id, enable_pag=True, torch_dtype=torch.float16)
         pipeline.enable_model_cpu_offload(device=torch_device)
         pipeline.set_progress_bar_config(disable=None)
@@ -319,6 +320,8 @@ class StableDiffusionXLPAGImg2ImgPipelineIntegrationTests(unittest.TestCase):
         )
 
     def test_pag_uncond(self):
+        from diffusers import AutoPipelineForImage2Image
+
         pipeline = AutoPipelineForImage2Image.from_pretrained(self.repo_id, enable_pag=True, torch_dtype=torch.float16)
         pipeline.enable_model_cpu_offload(device=torch_device)
         pipeline.set_progress_bar_config(disable=None)
