+++ tests/pipelines/pag/test_pag_sdxl_inpaint.py
@@ -33,7 +33,6 @@ from transformers import (
 
 from diffusers import (
     AutoencoderKL,
-    AutoPipelineForInpainting,
     EulerDiscreteScheduler,
     StableDiffusionXLInpaintPipeline,
     StableDiffusionXLPAGInpaintPipeline,
@@ -308,6 +307,8 @@ class StableDiffusionXLPAGInpaintPipelineIntegrationTests(unittest.TestCase):
         return inputs
 
     def test_pag_cfg(self):
+        from diffusers import AutoPipelineForInpainting
+
         pipeline = AutoPipelineForInpainting.from_pretrained(self.repo_id, enable_pag=True, torch_dtype=torch.float16)
         pipeline.enable_model_cpu_offload(device=torch_device)
         pipeline.set_progress_bar_config(disable=None)
@@ -325,6 +326,8 @@ class StableDiffusionXLPAGInpaintPipelineIntegrationTests(unittest.TestCase):
         )
 
     def test_pag_uncond(self):
+        from diffusers import AutoPipelineForInpainting
+
         pipeline = AutoPipelineForInpainting.from_pretrained(self.repo_id, enable_pag=True, torch_dtype=torch.float16)
         pipeline.enable_model_cpu_offload(device=torch_device)
         pipeline.set_progress_bar_config(disable=None)
